+++ CHANGELOG.md
@@ -1,3 +1,26 @@
+# [0.0.4]
+## Added:
+- Add `Entry` Properties.
+- Add `WindowPosition`.
+- Add All `WindowPosition` Properties.
+- Add `set_position`.
+- Add `set_halign` & `set_valign` & `set_sensitive`.
+- Add `halign` & `valign` & `sensitive` args.
+- Add Advanced import suggests with `Gtk` & `Gdk` & `GdkPixbuf` i'll add more later.
+- Add `set_logo`.
+- Add `pixbuf = GdkPixbuf.Pixbuf.new_from_file("logo.png")` Auto Complete.
+- Add Auto Suggests Pics Paths With GdkPixbuf Method `new_from_file()`.
+- Add `set_margen_start` & `set_margin_end` & `set_margin_top` & `set_margin_bottom`.
+## Fixed:
+- Fix `Gtk.Window.[no suggest __init__()]`.
+- Fix `__init__([no suggest {label=""}])`
+- Fix Random Spacing Suggests.
+- Fix `Align` Properties.
+## Special:
+- Finished my Exam i can focus now on extension.
+- Finaly Added The Sadly `Entry` :)
+- small random fix for UX if type `show_all()` text cursor don't go to in the Parentheses.
+
 # [0.0.3]
 ## Added:
 - Add `Gtk.Entry` and `Gtk.AboutDialog` Widgets.